refactor: rename types > index and cleanup imports and main index
