+++ types/search/index.ts
