+++ types/stamboek/index.ts
